api_gateway/requirements.txt
Django==5.2.1
djangorestframework==3.15.0
drf-yasg==1.21.7
requests==2.31.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
gunicorn 